# perf(backend): avoid N+1 selects from `user` table when packing many entities