… packages/backend/src/core/entities/ReversiGameEntityService.ts
… packages/backend/src/server/api/endpoints/notes/reactions.ts
… packages/backend/src/server/api/endpoints/roles/users.ts
… packages/backend/src/server/api/endpoints/users/get-frequently-replied-users.ts
… packages/backend/src/server/api/endpoints/users/show.ts