<!-- ℹ お読みください / README
PRありがとうございます！ PRを作成する前に、コントリビューションガイドをご確認ください:
Thank you for your PR! Before creating a PR, please check the contribution guide:
https://github.com/misskey-dev/misskey/blob/develop/CONTRIBUTING.md
-->

## What
<!-- このPRで何をしたのか？ どう変わるのか？ -->
<!-- What did you do with this PR? How will it change things? -->
#13910 に書いたことを実施しました。

## Why
<!-- なぜそうするのか？ どういう意図なのか？ 何が困っているのか？ -->
<!-- Why do you do it? What are your intentions? What is the problem? -->
Closes #13910

## Additional info (optional)
<!-- テスト観点など -->
<!-- Test perspective, etc -->
とりあえず https://gist.github.com/zyoshoka/647b9e9b74313e04c80a2230e014b088 のように書いたパフォーマンス計測用コードを `packages/backend/test/e2e` 直下に置いて E2E テストを実行すると次のようになりました。

### Before
```
admin/abuse-user-reports: 545.7474999999977 ms
blocking/list: 508.15108300000065 ms
users/show: 436.0121250000011 ms
```

### After
```
admin/abuse-user-reports: 49.55412500000239 ms
blocking/list: 36.62649999999849 ms
users/show: 27.77375000000029 ms
```

## Checklist
- [x] Read the [contribution guide](https://github.com/misskey-dev/misskey/blob/develop/CONTRIBUTING.md)
- [ ] Test working in a local environment
- [ ] (If needed) Add story of storybook
- [ ] (If needed) Update CHANGELOG.md
- [ ] (If possible) Add tests